+++ Sources/DeskPins/WindowMirror.swift
@@ -276,7 +276,6 @@ enum ScreenCapturePermission {
     }
 
     static func openSettings() {
-        let url = URL(string: "x-apple.systempreferences:com.apple.preference.security?Privacy_ScreenCapture")!
-        NSWorkspace.shared.open(url)
+        PrivacySettings.open(.screenRecording)
     }
 }